attempt at using webworkers
it was slower overall. not quite sure why, but it looks like webworkers
weren't really suited for this real time use case (at least until
SharedArrayBuffer gains traction).

Thoughts:

1. having to copy the data every frame is a killer (actually maybe
explore sending JSON or other array data)
2. I couldn't send the uint8 fgmaskData over for some reason.
3. even more perf - write custom shader
4. it didn't look much better. The main thread *was* at a nice 60fps
(<16ms/frame) but we'd have frames where "nothing happens" because the
worker takes 30ms to compute + send the data over, which effectively
means it looks the same as the singlethreaded version
5. oh maybe i should try spawning multiple workers? let me try that

* send JSON over/back
* debug fgmaskData
* use multiple workers
